chore(uppercase): move `uppercase` types to `uppercase` directory.
