+++ src/lib/uppercase/lib/index.ts
@@ -1,0 +1,2 @@
+export type { UppercaseLetterRangePattern } from './uppercase-letter-range-pattern.type';
+export type { UppercaseLetterRange } from './uppercase-letter-range.type';
